add codesnip_attr
