+++ crates/codesnip/Cargo.toml
@@ -4,10 +4,6 @@ version = "0.1.0"
 authors = ["to-omer <toomerhs@gmail.com>"]
 edition = "2018"
 
-[lib]
-doc = false
-proc-macro = true
-
 [[bin]]
 name = "cargo-codesnip"
 path = "src/main.rs"
@@ -15,9 +11,8 @@ doc = false
 
 [dependencies]
 codesnip_core = { path = "../codesnip_core" }
+codesnip_attr = { path = "../codesnip_attr" }
 syn = { version = "1", features = ["full"] }
-proc-macro2 = "1"
-quote = "1"
 structopt = "0.3"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
