+++ crates/codesnip_attr/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "codesnip_attr"
+version = "0.1.0"
+authors = ["to-omer <toomerhs@gmail.com>"]
+edition = "2018"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+codesnip_core = { path = "../codesnip_core" }
+syn = { version = "1", features = ["full"] }
+proc-macro2 = "1"
+quote = "1"
